backend/requirements.txt
fastapi
httpx
pydantic
python-dotenv
chromadb
langchain
langchain-community
pdfplumber
tempfile
